Moved implementation of IdexEnum into enum_index_derive.
