--- index_enum_derive/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "index_enum_derive"
-version = "0.1.0"
-authors = ["Omar Zeidan <omar@zeidan.io>"]
-license = "MIT"
-
-description = "Macros for accessing enum variant via index"
-keywords = ["enum", "macros", "index"]
-
-readme = "../README.md"
-
-[dependencies]
-syn = "0.11.11"
-quote = "0.3.15"
-
-[lib]
-proc-macro = true
